pyproject.toml
[project]
name = "whisper-openvino"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "huggingface-hub[hf-xet]>=0.36.0",
    "librosa>=0.11.0",
    "openvino-genai>=2025.3.0.0",
    "optimum-intel[openvino]>=1.26.1",
    "transformers==4.45",
]
